Fixed board creation
Now will return a command list to create a board that has a solution.
Flooding is performed on every board to get from the start node to the
goal node.

NOTE: finding the wumpus is never certain on any of these boards since
it is worth no points. If we can never get to a wumpus square, that is
fine.